requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
WTForms==3.0.1
python-dotenv~=0.20
stytch~=7.6
pymongo~=4.5.0
email-validator==2.0.0.post2
